+++ kickIn_26/src/main/java/nl/utwente/di/team26/Utils.java
@@ -4,8 +4,8 @@ import javax.ws.rs.core.SecurityContext;
 
 public class Utils {
 
-    public static int getUserFromContext(SecurityContext securityContext) {
-        return Integer.parseInt(securityContext.getUserPrincipal().getName());
+    public static long getUserFromContext(SecurityContext securityContext) {
+        return Long.parseLong(securityContext.getUserPrincipal().getName());
     }
 
 }
